launcher=Dir/ error handling & code comments.
Consistently with the submenu= keyword semantics, now launcher=Dir/
builds a sub-menu of MAX_SUBMENU_DEPTH levels. If .desktop files are
located in inferior sub-levels, the warning message "max submenu depth
exceeded for <inferior-branch>" is printed, and an empty gtkmenu leaf
menu is added to the menu path that leads to <inferior-branch>, for all
branches whose .desktop files can't be added to the menu.
This behavior might change in the future, as the empty leaf menu looks
ugly.