Format="single" doesn't work as advertised (no plan to fix it).
I couldn't find anything in the legacy code that supports the
statement
 If "formatting" contains only one "format_string", that "format_string"
 applies to everything following, no matter where it is in the menu
 hierarchy."
So I marked it as invalid. I have no plan to extend the legacy code to
support that statement, because sub-menu levels can be formatted
differtly by specifying a submenu= keyword at each level.

Note that formatting launcher=dir sub-levels is a different story,
since there is no spot in the configuration file where format= can be
added for sub-directories. So my plan is to  add that to the
specification of the .gtkmenuplus.desktop file.