Change activation log file format and policy.
It is formatted as a menu configuration file for easy inclusion in
other menus: Meta+Item+Cmd+Icon[+Tooltip]+EndData.
Meta :=
  '#{'
    <n>
    ':' <activation_count>
    ':' <creation_date>
    ':' <activation_date>
  '}'
  [ <user_data> ]
  '\n'
EndMeta ::=
  '#{}'
  [ <user_data> ]
  '\n'
The user shall never change the file outside the optional <user_data>
entries.  New items are appended to the file.  For already-logged
items Gtkdialogplus updates <activation_count> and <activation_time>.
Gtkdialogplus never rearranges or removes lines from the activation log
file. It is up to the calling application to do so.