Allow very long comment lines in scripts.[1]
[1] When the internal comment-gathering feature is turned off.
Before this change the legacy code printed a criptic error message
'no = sign' when a comment was longer than MAX_LINE_LENGTH characters.
Now it just skips the comment line, regardless of its length.
I had to change the default comment-gathering value to FALSE - I'm not
sure why it was TRUE. The change is hardcoded, but I will soon add a
command-line switch to enable it.