Remove launcher= restriction on symlinks.
Before this change the argument to launcher= had to end with ".desktop"
for the launcher to be found. So launcher=file didn't recognize "file"
as a valid desktop file.
After this change you can write launcher=link and make link point to
link -> /path/a -> /path/b/file.desktop. So "link" is recognized as a
valid desktop file if it ultimately points to a file whose name ends
with ".desktop".