Fix legacy bug: cmd=doc displays error icon (note)
According to menu_configuration_file_format.txt `cmd=document-path`
should display the icon of the application associated with
document-path. Before this fix that wasn't the case only when the
application icon was an icon name, such as mtpaint, but not when it was
a icon filepath.
See also test case script tests/noicons.gtkmenuplus.