report error if passed a directory as input
Test case: ` gtkmenuplus-debug /tmp `.

Also report an error if argv[1] is a relative path not starting with '.'.
Before this commit the path was handled as a variable name.

Test case: ` gtkmenuplus-debug test/abspath.gtkmenuplus `.

NOT CHANGED: if argv[1] is a directory or file name ending with `=` in the
current directory, argv[1] is treated as a file and not as a variable assignment.

Test case: ` (cd /tmp && touch 'z=' && gtkmenuplus-debug test z=) `.

> _Reads file `/tmp/z=`._